novena-disable-ssp: Add an overly complex program to disable SSP
SSP is a "simple secure pairing" feature that some dongles support.
Unfortunately, this seems to break reliable pairing for keyboards, so
it makes sense to disable it.

The code could stand to be a lot cleaner.  It is currently based off
btmon, which is needlessly complex.  It would be nice to pare it down.